# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Sistema completo de gestión de cuestionarios NOM-035 (y otros dominios) para empresas. Soporta múltiples tipos de preguntas, ciclo de vida de cuestionarios por empresa con máquina de estados, asignación jerárquica por roles, reportes agregados y exportación de datos.",
    "title": "Questionnaire Service API - NOM-035 / Wemoova",
    "contact": {
      "name": "Wemoova Technologies",
      "email": "support@wemoova.com"
    },
    "version": "1.1"
  },
  "host": "qa.services.wemoova.com",
  "basePath": "/questionarie-service",
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer {token}' to authorize. Get token from FusionAuth login."
    }
  },
  "paths": {
...
    "/api/v1/questionnaires": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List all questionnaires (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "List Questionnaires",
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "page_size",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "active",
            "in": "query",
            "description": "Filter by active only"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create a new questionnaire (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "Create Questionnaire",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "title"
              ],
              "properties": {
                "title": {
                  "type": "string",
                  "example": "NOM-035 Guía III 2025"
                },
                "description": {
                  "type": "string"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "nom-035",
                    "riesgo-psicosocial"
                  ]
                },
                "category_id": {
                  "type": "string",
                  "description": "ID de categoría (opcional)"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/questionnaires/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get questionnaire by ID (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "Get Questionnaire",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update questionnaire (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "Update Questionnaire",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "category_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Deactivate questionnaire (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "Deactivate Questionnaire",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated"
          }
        }
      }
    },
    "/api/v1/questionnaires/{id}/stats": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get stats for a specific questionnaire: total questions, required questions, tags, category (Super Admin only)",
        "tags": [
          "Questionnaires"
        ],
        "summary": "Get Questionnaire Stats",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "questionnaire_id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "total_questions": {
                  "type": "integer"
                },
                "required_questions": {
                  "type": "integer"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "category_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questionnaires/{id}/questions": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add a question to questionnaire (Super Admin only)",
        "tags": [
          "Questions"
        ],
        "summary": "Add Question",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "question_text",
                "question_type"
              ],
              "properties": {
                "question_text": {
                  "type": "string",
                  "example": "¿Mi trabajo me permite desarrollar nuevas habilidades?"
                },
                "question_type": {
                  "type": "string",
                  "enum": [
                    "multiple_choice",
                    "likert_scale",
                    "free_text",
                    "yes_no"
                  ]
                },
                "options": {
                  "type": "object",
                  "example": {
                    "min": 1,
                    "max": 5,
                    "labels": {
                      "1": "Nunca",
                      "5": "Siempre"
                    }
                  }
                },
                "is_required": {
                  "type": "boolean",
                  "example": true
                },
                "order_index": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/questionnaires/{id}/questions/{question_id}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update a question (Super Admin only)",
        "tags": [
          "Questions"
        ],
        "summary": "Update Question",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "question_id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Remove a question (Super Admin only)",
        "tags": [
          "Questions"
        ],
        "summary": "Remove Question",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "question_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Removed"
          }
        }
      }
    },
    "/api/v1/questionnaire-categories": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List all questionnaire categories (Super Admin only)",
        "tags": [
          "Categories"
        ],
        "summary": "List Categories",
        "parameters": [
          {
            "type": "boolean",
            "name": "active",
            "in": "query",
            "description": "Filter active only"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create a questionnaire category (Super Admin only). Useful for grouping: NOM-035, Bienestar, Clima Laboral, Onboarding.",
        "tags": [
          "Categories"
        ],
        "summary": "Create Category",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "NOM-035"
                },
                "description": {
                  "type": "string",
                  "example": "Norma Mexicana de Riesgos Psicosociales"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/questionnaire-categories/{id}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update a category (Super Admin only)",
        "tags": [
          "Categories"
        ],
        "summary": "Update Category",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get a questionnaire category by ID (Super Admin only)",
        "tags": [
          "Categories"
        ],
        "summary": "Get Category by ID",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/questionnaire-categories/{id}/questionnaires": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List questionnaires belonging to a category (Super Admin only)",
        "tags": [
          "Categories"
        ],
        "summary": "Get Questionnaires by Category",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List all companies (Super Admin only)",
        "tags": [
          "Companies"
        ],
        "summary": "List Companies",
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create a new company (Super Admin only)",
        "tags": [
          "Companies"
        ],
        "summary": "Create Company",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Salud Mental MX S.A."
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/companies/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get company by ID (Super Admin only)",
        "tags": [
          "Companies"
        ],
        "summary": "Get Company",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update company (Super Admin only)",
        "tags": [
          "Companies"
        ],
        "summary": "Update Company",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "is_active": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Deactivate company (soft delete - Super Admin only)",
        "tags": [
          "Companies"
        ],
        "summary": "Deactivate Company",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/questionnaires": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List questionnaires assigned to a company (Company Admin+)",
        "tags": [
          "Company Questionnaires"
        ],
        "summary": "Get Company Questionnaires",
        "parameters": [
          {
            "type": "string",
            "name": "company_id",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "name": "active",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Assign questionnaire to company with period (Super Admin only). Creates a CompanyQuestionnaire in 'active' status.",
        "tags": [
          "Company Questionnaires"
        ],
        "summary": "Assign Questionnaire to Company",
        "parameters": [
          {
            "type": "string",
            "name": "company_id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "questionnaire_id",
                "period_start",
                "period_end"
              ],
              "properties": {
                "questionnaire_id": {
                  "type": "string"
                },
                "period_start": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-01-01"
                },
                "period_end": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-03-31"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request - period overlap or questionnaire inactive"
          }
        }
      }
    },
...
    "/api/v1/company-questionnaires/{id}/activate": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Activate a CompanyQuestionnaire (draft→active or paused→active). Company Admin+",
        "tags": [
          "Company Questionnaires"
        ],
        "summary": "Activate Company Questionnaire",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activated"
          },
          "400": {
            "description": "Invalid transition"
          }
        }
      }
    },
...
    "/api/v1/company-questionnaires/{cq_id}/pending-users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "List assignments with status 'pending' (not started) for a company questionnaire (Supervisor+)",
        "tags": [
          "Company Questionnaires"
        ],
        "summary": "Get Pending Users",
        "parameters": [
          {
            "type": "string",
            "name": "cq_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}